+++ output.yml
@@ -1,0 +1,13 @@
+apiVersion: v1
+kind: query
+metadata:
+  name: "user_login_activity_last_24_hours"
+spec:
+  name: osquery_info
+  query: |
+    SELECT username, host, time, tty, from_host
+    FROM last
+    WHERE time > (datetime('now', '-24 hours'));
+  description: "This query returns all login activity in the past 24 hours."
+  platform: linux
+  interval: 3600
