+++ output_fleet.json
@@ -1,0 +1,20 @@
+[
+  {
+    "name": "schedule",
+    "query": "select name, interval, executions, output_size, wall_time, (user_time/executions) as avg_user_time, (system_time/executions) as avg_system_time, average_memory, last_executed, denylisted from osquery_schedule;",
+    "description": "Imported from osquery pack. Interval: 7200",
+    "platform": "all"
+  },
+  {
+    "name": "events",
+    "query": "select name, publisher, type, subscriptions, events, active from osquery_events;",
+    "description": "Imported from osquery pack. Interval: 86400",
+    "platform": "all"
+  },
+  {
+    "name": "osquery_info",
+    "query": "select i.*, p.resident_size, p.user_time, p.system_time, time.minutes as counter from osquery_info i, processes p, time where p.pid = i.pid;",
+    "description": "Imported from osquery pack. Interval: 600",
+    "platform": "all"
+  }
+]
